+++ src/main/java/inventorymanagement/controller/ProductController.java
@@ -2,6 +2,8 @@ package inventorymanagement.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,6 +15,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import inti.ws.spring.exception.client.UnauthorizedException;
+import inventorymanagement.constants.Constants;
 import inventorymanagement.model.IncomingProductModel;
 import inventorymanagement.model.ProductModel;
 import inventorymanagement.service.ProductServiceInterface;
@@ -28,7 +32,11 @@ public class ProductController {
   @RequestMapping(value = "/products", method = RequestMethod.POST)
   @ResponseBody
   @ResponseStatus(HttpStatus.CREATED)
-  public ProductModel create(@RequestBody IncomingProductModel productModel) {
+  public ProductModel create(@RequestBody IncomingProductModel productModel, HttpSession session) throws UnauthorizedException {
+    Boolean admin = (Boolean) session.getAttribute(Constants.SESSION_ADMIN);
+    if (admin == false) {
+      throw new UnauthorizedException("Unauthorized access");
+    }
     LOG.info("Request received for adding a product");
     ProductModel product = productService.addProduct(productModel);
     LOG.info("Request for adding a product successful");
@@ -38,7 +46,11 @@ public class ProductController {
   @RequestMapping(value = "/products/{id}", method = RequestMethod.PATCH)
   @ResponseBody
   @ResponseStatus(HttpStatus.OK)
-  public ProductModel update(@PathVariable int id, @RequestBody IncomingProductModel productModel) {
+  public ProductModel update(@PathVariable int id, @RequestBody IncomingProductModel productModel, HttpSession session) throws UnauthorizedException {
+    Boolean admin = (Boolean) session.getAttribute(Constants.SESSION_ADMIN);
+    if (admin == false) {
+      throw new UnauthorizedException("Unauthorized access");
+    }
     LOG.info("Request received for updating a product");
     ProductModel product = productService.updateProduct(id, productModel);
     LOG.info("Request for updating a product successful");
@@ -48,7 +60,11 @@ public class ProductController {
   @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
   @ResponseBody
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void delete(@PathVariable int id) {
+  public void delete(@PathVariable int id, HttpSession session) throws UnauthorizedException {
+    Boolean admin = (Boolean) session.getAttribute(Constants.SESSION_ADMIN);
+    if (admin == false) {
+      throw new UnauthorizedException("Unauthorized access");
+    }
     LOG.info("Request received for deleting a product");
     productService.deleteProduct(id);
     LOG.info("Request for deleting a product successful");
